+++ src/models/ColumnsModel.ts
@@ -58,7 +58,9 @@ export function myBillsColumns () {
       sortable: true,
       align: 'left'
     },
-    { name: 'dueDate', field: 'dueDate', label: i18n.global.t('columns.bills.dueDate'), sortable: true, align: 'left' }
+    { name: 'dueDate', field: 'dueDate', label: i18n.global.t('columns.bills.dueDate'), sortable: true, align: 'left' },
+    { name: 'value', field: 'value', label: i18n.global.t('columns.bills.value'), sortable: true, format: val => `${i18n.global.t('generic.currencySymbol')} ${val}`, align: 'left' },
+    { name: 'actions', field: 'action', label: i18n.global.t('columns.bills.actions'), sortable: false, align: 'right' }
   ]
 }